+++ Pemesanan-Tiket-Pesawat/T_Economy.java
@@ -7,14 +7,14 @@ import java.util.Scanner;
  * @version 1.0
  */
 public class T_Economy extends Tiket {
-    private double totalRun = 0;
+    private int totalRun = 0;
     Scanner in = new Scanner(System.in);
 
     /**
      * Constructor untuk objek t_Economy tanpa paramter
      */
     public T_Economy() {
-        super(1000000.0);
+        super(1000000);
     }
 
     /**
@@ -22,7 +22,7 @@ public class T_Economy extends Tiket {
      * 
      * @param harga
      */
-    public T_Economy(Double harga) {
+    public T_Economy(int harga) {
         super(harga);
     }
 
@@ -41,7 +41,7 @@ public class T_Economy extends Tiket {
      * 
      * @return total harga tiket
      */
-    public double getTotalRun() {
+    public int getTotalRun() {
         return this.totalRun;
     }
 
@@ -50,7 +50,7 @@ public class T_Economy extends Tiket {
      * 
      * @param Subtotal
      */
-    public void setTotalRun(double Subtotal) {
+    public void setTotalRun(int Subtotal) {
         this.totalRun = Subtotal;
     }
 
@@ -60,8 +60,8 @@ public class T_Economy extends Tiket {
      * @param harga
      * @param jumlah
      */
-    public double getTotal(double harga, int jumlah) {
-        double totalSementara = (harga * jumlah);
+    public int getTotal(int harga, int jumlah) {
+        int totalSementara = (harga * jumlah);
         System.out.println("Subtotal: " + totalSementara);
         totalRun += totalSementara;
         return totalSementara;
@@ -73,8 +73,8 @@ public class T_Economy extends Tiket {
      * @param jumlah
      * @return
      */
-    public double deletePesanan(int jumlah) {
-        double temp = (getHargaTiket() * jumlah);
+    public int deletePesanan(int jumlah) {
+        int temp = (getHargaTiket() * jumlah);
         return temp;
     }
 }
